+++ src/onegov.core/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 Unreleased
 ~~~~~~~~~~
 
+- Caches the translator on the request to be slightly more efficient.
+  [href]
+
 0.0.1 (2015-04-29)
 ~~~~~~~~~~~~~~~~~~~